requirements.txt
Flask==1.0.2

# Application server for both development and production.
gunicorn==19.9.0

# Testing and static analysis.
pytest==2.9.1
pytest-cov==2.2.1
flake8==2.5.4

# CLI.
Click==7.0

# Data and workers.
redis==2.10.5
celery==3.1.23

# Forms.
#Flask-WTF==0.9.5
Flask-WTF==0.14.2
WTForms-Components==0.9.7

# Extensions.
flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1

absl-py==0.7.0
astor==0.7.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.16.0
opencv-python==3.2.0.8
Pillow==5.4.1
protobuf==3.6.1
PyYAML==3.13
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug


# db requirements
flask-sqlalchemy==2.3.2
Flask-Table==0.5.0

#getdominantcolour
colorthief==0.2.1